Move indexing into single module
Crated a module named indexing in position crate and moved indexing enums there.
