+++ position/src/lib.rs
@@ -5,6 +5,7 @@
 */
 pub mod bitboard;
 pub mod board;
+pub mod indexing;
 pub mod move_generation;
 pub mod position;
 pub mod prelude;
