+++ position/src/position/zobrist.rs
@@ -1,5 +1,4 @@
-use crate::bitboard::{File, Square};
-use crate::board::{Color, Piece};
+use crate::indexing::{Color, File, Piece, Square};
 use crate::position::CastlingRights;
 
 use const_random::const_random;
