+++ position/src/prelude.rs
@@ -1,4 +1,5 @@
-pub use crate::bitboard::{BitBoard, File, NegDiag, PosDiag, Rank, Square};
-pub use crate::board::{Board, Color, Piece};
+pub use crate::bitboard::BitBoard;
+pub use crate::board::Board;
+pub use crate::indexing::{Color, File, NegDiag, Piece, PosDiag, Rank, Square};
 pub use crate::move_generation::{ChessMove, ChessMoveHint, ChessUnmove};
 pub use crate::position::{CastlingRights, ParseFenError, Position, PositionHash};
